Deprecate legacy ctor and setResourceId method in Image.Builder API
This ctor and methods shouldn't be mixed with the new API that has
ProtoLayoutScope and automatic resource registration. To avoid
confusion with developers, we are deprecating the existing ones
as going forward, new evolved API should be used as it simplifies
developers usage and removes the need for `onTileResourcesRequest`
override.

Fixes: 432758526
Test: N/A
Relnote: "We have deprecated `Image.Builder()` and `Image.Builder.setResourceId` methods in favour of the new automatic resource registration API, available in `Image.Builder(ProtoLayoutScope)` and `Image.Builder.setImageResource` APIs that remove the need for overriding `onTileResourcesRequest`."
Change-Id: I7bfe6ecc4e5849a27eda55dd993273e440ba69ae