Refactor KspAnnotated.UseSiteFilter to be an enum.
The `UseSiteFilter` is used like an enum, so might as well make it one.
This change makes things a bit cleaner at the usage and declaration
site.

This CL also reorders the calls in the `apply` method so that it doesn't
try to eagerly call `getDeclaredTargets` which can be expensive.

Test: This CL is a noop change so relies on existing tests.
Change-Id: Ifa0a51e7546c1471978050d3c6508961d099f8d6