Fix for incorrect/unexpected onPlaced/onLayoutRectChanged callbacks during alignment lines
We were setting isPlaced and calling onPlaced callbacks with incorrect coordinates during the alignment calculation placement pass. It is wrong, as the node should only be considered placed when placeable.place() is called for this node and all of its parents.

Previous attempt to merge it caused b/436622681. The issue from that was fixed and covered with a new test PlacedChildTest.addingChildWithBaselineLater_isDisplayedAndOnPlacedIsCalled.

Test: PlacedChildTest, OnGlobalRectChangedTest

Bug: 436622681
Bug: 434151773
Bug: 433494436
Change-Id: I1e83e42a7f9bdad87c2669ec9a0c2a739b51aa52