Refactor DocumentSelection to support multiple selection types.
This change updates the `DocumentSelection` data class to handle different types of selections beyond a single continuous one. It introduces the ability to manage multiple, non-contiguous selections on a single page.

This enhanced flexibility is a foundational step for future features that will allow users to select and interact with multiple distinct text blocks or regions simultaneously.

Bug: 440474766
Test: ./gradlew :pdf:integration-tests:connectedAndroidTest
Change-Id: Iab806c1577da46229749fed578bf99b81504ceeb