Re-enable some of the Ignored tests in the runtime.
Re-enabled a few of the tests reported in b/438912113.

There is one still remaining but it uncovered that the bug, b/346821372
was not fixed by the compiler change. The compiler is now correct but
the test still fails as described in a new comment in the bug.

Excluding the one above, the remaining ignored tests are ignored
for valid platform reasons and need to remain ignored until they
are supported by the targets in question.

Relnote: N/A - Test clean-up
Bug: 438912113
Test: ./gradlew :compose:ru:r:tA
Change-Id: Ib4d705074bd59d46e783caea2405df69a73a402c