Use new initial anchor when pane expansion anchor list changes
Originally when the pane expansion anchor list changes, and the current
anchor is no longer in the list, we will just reset the current anchor
to null. This CL changes the behavior so that when the anchor list
changes, we will get the new initial anchor from the new list according
to the given intital anchor index, if any.

Relnote: Use new initial anchor when pane expansion anchor list changes
Test: instrumentation test
Fixes: 438829477
Change-Id: I91cd1ecdcc587722214757e05c8b33a6ef70a43b