Add equals() and hashCode() to composition data groups
Implemenattions of androidx.compose.runtime.tooling.CompositionGroup
now support `equals()` and `hashCode()` and will compare `==` if they
refer to the same group in the same table. Comparison for `==` is only
valid for the same slot table versiond. That is, if the slot table
change all `CompoistionGroup` produced for the table should be
considered invalid and will not compare equal to each other even if
they refer to the same group. Tracking a group when changes are made
to the table still requires `identity`.

Fixes: 436866433
Relnote: N/A
Test: ./gradlew :compose:ru:r:tA
Change-Id: I62d8231d9f3dec221b3990405651457c4d5b38c3