Downgrade CameraExecutor rejection log to warning
Rejected executions in CameraExecutor are currently
logged as errors. This situation often occurs when
CameraX#shutdown() is called, as some in-flight signals
from the camera may still attempt to execute tasks on
the executor that is in the process of shutting down.

Since these rejections are an expected part of the
shutdown sequence and the tasks can be safely ignored,
logging them as errors is unnecessarily alarming. This
change downgrades the log level from ERROR to WARNING
to reduce log noise.

Bug: 429278292
Test: ./gradlew bOS
Change-Id: Icc31ac724fc2b88f21498d9ec7f6de37d6b4fb94