Refactor UpdateSelectionAsync to support custom logic.
The `UpdateSelectionAsync` method is being refactored to be more generic. This change introduces a new `updateSelection` lambda parameter, allowing the caller to provide custom selection logic while reusing the core asynchronous update and rendering flow.

This preparatory refactoring simplifies the codebase and makes it easier to implement more complex selection behaviors in the future.

Bug: 441199935
Test: ./gradlew :pdf:integration-tests:connectedAndroidTest
Change-Id: I12e95bcf575fcb16c8103a00e64cf0e6efc1dc2c