Take into account layers transformations from outer coordinator
When the parent layout is placing a child with layer transformation, we weren't applying it.
Plus, the matrix transformations were incorrectly applyed when we both have layers modifications and an offset. We should first apply the matrix, and only then translate by the offset when we go up the hierarchy. Previously we were doing it in a wrong order.

Test: new tests
Bug: 434921699
Change-Id: Ia05ac9e65e6e5608f1433eb275e9706097db1b11