Refactor ComposerImpl into its own file.
This is the first step in refactoring the current composer
into a state where we can a new implementation of the composer,
called LinkComposer, behind a flag.

This change move the current composer, ComposerImpl, into
its own file, ComposerImpl.kt, leaving the public API in
Composer.kt.

Test: N/A - refactor
Relnote: N/A
Bug: 268366116
Change-Id: I0b56ef4e4c12b64fcb375d17b996cccdb506127b