Intercept long-press events on hyperlinks in SelectionStateManager
This change introduces handling for long-press gestures on embedded hyperlinks within PDF documents.

Previously, long-pressing on a link would trigger the standard text selection mode. This change intercepts the long-press event when it occurs on a link. Instead of initiating text selection, it prepares for a link-specific action mode. This will enable a more intuitive user experience by allowing for future work to display a context menu with link-related actions (e.g., open, copy).

The private scope for SelectionModel was changed to internal as Backend models for LinkSelection do not support linkText which is a crucial part of frontend selection model. Thus we need to construct it on our own.

This is a foundational step towards richer link interaction in the PDF viewer.

Bug: 436991645
Test: ./gradlew :pdf:integration-tests:connectedAndroidTest
Change-Id: I8d35cb8ab1df5798ca61e399c9c344558f0c1a38