Makes the measure policy define how much scroll it consumed and accumulated
With Adaptive scrolling, we need more control over what the Glimmer List reports as consumed and accumulates for the next pass.
This change prepares for that by moving the calculation of reported and accumulated values from ListState to MeasurePolicy. The logic remains the same so far; only the location has changed.

Test: Checked that the old tests pass and added new ones to the ListStateTest
Change-Id: I7f87423e4db6070a28b968bbbf21c8c171e53aeb