Simplify Kotlin project detection in SafeArgs plugin
The SafeArgs Gradle plugin now detects if a project is a Kotlin project by checking for the presence of the `kotlin` extension.

This replaces the previous implementation that iterated over a hardcoded list of Kotlin plugin IDs using `project.plugins.withId`. This new approach also handles builtin kotlin

BUG: 439568035
TESTED: test command passed for that module
Change-Id: I1d9d76416ff68773269b7825cf912c373f044bec