Implement rememberDecoratedNavEntries
The function converts a backStack List<Any> into a list of decorated NavEntries. This list of entries is remembered and updates when there are changes to the passed in backStack (add to, remove from, and replace stack) or entry decorators.

Bug: 441328236
Test: ./gradlew navigation3:navigation3-runtime:androidConnectedCheck
Relnote: "rememberDecoratedNavEntries has replaced DecoratedNavEntryProvider to only create and return NavEntries decorated with the list of decorators"
Change-Id: I0fe1ce4618e337102a3daa01b5bca94a592dd6f8