Split out the initialization process of the UseCaseCamera component
Currently, when the UseCaseCamera component is built, the internal
surface manager is initialized as the Dagger components are built.
However, the surface manager initialization process can trigger use case
reset, which may cause cyclic Dagger injections as it refreshes the use
cases and creates a new UseCaseCamera component.

This CL splits/separates out the initialization (and start) process of
the UseCaseCamera component, so a cyclic injection scenario can never
happen again.

Bug: 434382869
Test: Run affected tests
Change-Id: I8b4d5196ec7f78e8b88beb7871031fcd02edba17