Handle potential camera errors immediately following camera open
When opening cameras, it is possible for the camera to be closed due to
an error. This may prevent us from acquiring a WakeLock token which we
use to ensure the availability of the (real) camera till it's connected
to a VirtualCamera.

Handle this by checking whether we succeed in acquiring the token, and
treate it as a failure if we fail to do so.

Bug: 443331065
Test: Unit tests
Change-Id: If525ef5f564b3c5fb04042e8f7bcd3a898ae3b44